Drop userspace (libmctp) solution
It turned out that the current 'kcs_bmc_cdev_raw' driver can't be used
as-is in the libmctp code. And the modifications for the driver are not
that simple and even if possible probably not worth the effort since
the Linux kernel driver approach is on the way.

For example in some points KCS state machine requires BMC to read STR,
but in the process the code can't read IDR. And the 'kcs_bmc_cdev_raw'
driver doesn't satisfy such requirements.

With these race cases communication is not reliable, sometimes your
packets would go flawlessly and sometimes KCS machine would stuck.

The error that I've caught was connected with the WR_END command byte:
HOST writes WR_END=0x62 command, but BMC sees 0x62 as DATA (bit 4 in
the STR register is read as 0).

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>